guest: find and use vcpu signature to calculate SNP launch digest
The launch digest should also include the vcpu_signature to match the
hardware measurement. This patch adds a vcpu_type parameter to the
LaunchDigest* routines, which is used to look up the CPU signature from
the existing CpuSigs table.

If vcpu_type is unspecified, it defaults to using a signature value of
0.

Signed-off-by: Jagannathan Raman <jag.raman@oracle.com>
Reviewed-by: Ian Chin Wang <ian.chin.wang@oracle.com>