golang: update module name and golang imports to use virtee packages
go.mod still refers to the module name as
"github.com/edgelesssys/sev-snp-measure-go". This causes an error
when adding this module as a dependency from another project. As such, we
need to update the module's name to "github.com/virtee/sev-snp-measure-go".

While we are at it, we could also update the Golang imports to refer to
the virtee repo

Signed-off-by: Jagannathan Raman <jag.raman@oracle.com>
Reviewed-by: Liam Merwick <liam.merwick@oracle.com>